pyproject.toml
[project]
name = "chandra-ocr"
version = "0.1.9"
description = "OCR model that converts documents to markdown, HTML, or JSON."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Datalab", email = "hi@datalab.to"}
]
keywords = ["ocr", "pdf", "markdown", "layout"]
dependencies = [
    "beautifulsoup4>=4.14.2",
    "click>=8.0.0",
    "filetype>=1.2.0",
    "flask>=3.0.0",
    "markdownify==1.1.0",
    "pillow>=10.2.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pypdfium2>=4.30.0",
    "python-dotenv>=1.1.1",
    "qwen-vl-utils>=0.0.14",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.1",
    "streamlit>=1.50.0",
    "accelerate>=1.11.0",
]

[build-system]
requires = ["setuptools>=61"]  # or "setuptools>=61", "flit-core", etc.
build-backend = "setuptools.build_meta"

[project.scripts]
chandra = "chandra.scripts.cli:main"
chandra_app = "chandra.scripts.run_app:main"
chandra_api = "chandra.scripts.run_api:main"
chandra_screenshot = "chandra.scripts.screenshot_app:main"

[tool.setuptools.packages.find]
include = ["chandra*"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]
